feat(fiche-patient): US-2632 composant présentational <PatientRecord> + contrat
Extrait le rendu de la fiche patient (en-tête + 4 onglets) dans un composant
présentational réutilisable `src/components/diabeo/patient/PatientRecord.tsx`,
piloté par un DTO (`PatientRecordData`) et agnostique de la source de données.
`PatientDetailClient` devient un adaptateur PAGE mince qui câble <PatientRecord>
et fournit le contrat de liens (`documentHref` → ?patientId=). Le composant ne
construit aucune URL portant l'id patient → prêt pour l'adaptateur drawer cTok
(US-2633). Zéro changement fonctionnel/visuel.

Tests : parité préservée (25) + 1 test verrouillant la délégation à documentHref.
tsc/eslint/build OK, 3439 tests verts.
