+++ tests/components/patient-detail-client.test.tsx
@@ -57,6 +57,7 @@ vi.mock("@/components/diabeo/CgmChart", () => ({
 }))
 
 import { PatientDetailClient, type PatientDetailData } from "@/app/(dashboard)/patients/[id]/PatientDetailClient"
+import { PatientRecord } from "@/components/diabeo/patient/PatientRecord"
 
 const baseData: PatientDetailData = {
   id: 42,
@@ -161,6 +162,19 @@ describe("PatientDetailClient (Phase 1)", () => {
     const dl = screen.getByText("Télécharger").closest("a")
     // doit porter le patientId pour la résolution de scope côté route (pro)
     expect(dl?.getAttribute("href")).toBe("/api/documents/7/download?patientId=42")
+  })
+
+  // US-2632 — <PatientRecord> ne construit pas l'URL lui-même : il délègue au
+  // contrat `documentHref` (le drawer fournira une variante `cTok`).
+  it("delegates the document link to the documentHref contract (drawer-ready)", () => {
+    render(
+      <PatientRecord
+        data={baseData}
+        documentHref={(id) => `/cTok/documents/${id}`}
+      />,
+    )
+    const dl = screen.getByText("Télécharger").closest("a")
+    expect(dl?.getAttribute("href")).toBe("/cTok/documents/7")
   })
 
   it("shows the empty Documents state when there are none", () => {
